Update to use 1.6 double health vars
WARNING: The current Bukkit dev builds(2794) seem to have errors
themselves!

+++ DvZ/src/me/andre111/dvz/Spellcontroller.java
@@ -116,7 +116,7 @@ public class Spellcontroller {
 	}
 	
 	public static void spellSuizide(Game game, Player player) {
-		player.damage(1000);
+		player.damage((double) 1000);
 	}
 	
 	//create the portalblocks
